Rename module to avoid naming conflict.
