+++ src/openscm_units/__init__.py
@@ -1,9 +1,8 @@
 """
 OpenSCM Units, units for use with simple climate modelling
 """
+from ._unit_registry import ScmUnitRegistry, unit_registry
 from ._version import get_versions
-from .unit_registry import unit_registry  # noqa: F401
-from .unit_registry import ScmUnitRegistry
 
 __version__ = get_versions()["version"]
 del get_versions
